finally landed on a rather portable non-invasive API
see usage example in `rewrite.rs/rewrite` docblock

The idea is that cfsqlite is used as a library alongside a normal sqlite connection.

If you want to do crr things, you call `cfsqlite` and as it to re-write your create & alter table
statements to their conflict-free variants. You then issue those sql queries instead.

Not as transparent as I would like but if we wanted a fully transparent API we're left with a few options:

1. a custom build of sqlite and thus re-creating bindings for all target languages
2. a wrapper around the sqlite connection which again requires re-creating language bindings
3. a virtual table extension which, itself, is not completely transparent. E.g., requires doing `alter tables` via `inserts` against the virtual table.
4. get a change merged into sqlite core. We need a proof of concept and adoption and a million other things before we can really even start down this path.