bump db version in triggers rather than commit hooks
This gives each row that is touched a unique version which allows us to chunk the replication process.

If a bunch of rows have the same version we would need to ensure all of those rows are sent and applied
before moving to the next set of rows.

This is because peers keep track of what updates they have from other peers via peer version numbers.
If a peer A sends version 10 to peer B then peer B will never ask A for anything prior to 10.

Also, I'm skeptical that tagging all rows that are in a tx with the same version actually accomplishes anything for us.

Once state merging starts to happen, any guarantess provided by that transaction are lost to the merge process.
todo: diagram this out.