go back to using a commit hook to bump db version
given the new clock format we can chunk clocks with the same timestamp via `rowid` on the clock table.

while this doesn't guarantee preseservation of transactional invariants on replication, it'll help by allowing us to replicate all rows that were in a transaction.

Caveat: if a transaction commits then another transaction commits that partially overlaps the first transaction, we lose a record of some of the rows and values committed by the first transaction.