start implementing changes since vtab. requires altering dbversion bumping
vtabs must return rowids which must be unique for every row.
Given that, dbversion must be unique for each touched row.

Given that, dbversion is now bumped on every modification (insert/update/delete) of a crr row
rather than only on transaction commits.

If we want to track transactions in the future we'll need to keep a `txid` counter instead.

To support knowing which rows and columns are in