register commit hooks to roll back cached db version
We need to do this so we refresh the version on the next tx.

We need to refresh on the next tx because other processes might be writing data too.

Presumably we can get smarter here by only rolling back if `https://www.sqlite.org/pragma.html#pragma_data_version` changed.

But this assumes other application code is not making use of that pragma. If they are they can get us into
a state where data_version doesn't return a change since it gets reset between calls to it.