return result of transaction callback fixes #104
we have async and sync interfaces to the DB given that in the browser it runs async
and in node it can run sync.

If a user is writing cross-platform code, they'll be awaiting synchronous returns.

This is fine until they pass a callback that actually does return a promise. In that case, that value
needs to be returned back up the stack so `await` works properly.