more thoughts on and implementation of move and insert_after trigger
This is taking a bit given we do not want to have to prepare statements for the common case of a non-conflicting insert.

Preparing statements is _slow_ so we want to do as much in the trigger definition as we can.

The other part is that we don't want to have to query our schema table every time we need to resolve a conflict
so I needed to figure out how to retain and pass all that information around.

Looks like we can remove the schema table entirely -- but we won't given we'll need it for migrations