is_crr utility to short-circuit crr re-creation
This is now needed given we perform a backfill task on upgrading a table to a CRR and/or
post schema modification of a table.

If the table already was a CRR and never had a schema modification, no need to re-run the
CRR tasks + backfill.

It is a common pattern in apps to do this at startup:

db.apply_schema(some_schema);

where `some_schema` is:

```sql
CREATE TABLE IF NOT EXISTS tbl ....;
SELECT crsql_as_crr('tbl');
...
```

We could technically also abstract this away in our auto-migrate facility.

I.e.,

```sql
SELECT crsql_auto_migrate(schema_version, schema_string);
```

where, if there is no schema_version delta, we skip application of the schema_string.