use db_version for col_version
This ensures that if two fields should always change together that they always will as when one loses a merge the other one will also lose the merge.

Hmm, except in the case where the clocks are equal and then we tie break on value. Would be better to tie break on peer id for consistency.